scipy-2023/requirements.txt
dataclasses_json
flytekitplugins-deck-standard
flytekitplugins-kfpytorch
flytekitplugins-pandera
flytekitplugins-spark
flytekitplugins-whylogs
flytekit>=1.7.0
grpcio==1.51.3
ipykernel
joblib
modin
palmerpenguins
pandas
pyspark<3.4.0
scikit-learn
torch
wheel
whylogs
